requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
django-bootstrap4==22.1
Django==4.0.4
djangorestframework==3.13.1
pip==21.1.2
psycopg2==2.9.3
pytz==2022.1
setuptools==57.0.0
soupsieve==2.3.2
sqlparse==0.4.2
tzdata==2022.1
wheel==0.36.2
